separate deadlock tests
